requirements.txt
deepgram-sdk
python-dotenv
python-magic
requests
